refactor(hover): split hover helpers into focused modules (#0000)
